+++ src/TestPrune.SqlHydra/SqlHydraAnalyzer.fs
@@ -182,5 +182,5 @@ type SqlHydraExtension(generatedModulePrefix: string) =
 
         member _.AnalyzeEdges (symbolStore: SymbolStore) (_changedFiles: string list) (_repoRoot: string) =
             SqlHydraExtension.extractFacts generatedModulePrefix symbolStore
-            |> SqlCoupling.buildEdges
+            |> SqlCoupling.buildEdges (symbolStore.GetTestMethodSymbolNames())
             |> List.map (fun d -> { d with Source = "sql-hydra" })